backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.2.0
idna==3.10
kombu==5.5.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
